Update SensorHelper.ino
Lots of things were included in this update and the program is mostly finished. Button was made very functional so it can detect long and short presses. A counter was created and overall the program works sturdily. I am planning on creating a standby mode to finish the application off and then finally case it and use it. :)